+++ packages/plugin-s3/src/services/delete-multiple.ts
@@ -1,12 +1,8 @@
 import { type S3Client, DeleteObjectsCommand } from "@aws-sdk/client-s3";
-import type { NodeJsClient } from "@smithy/types";
 import type { PluginOptions } from "../types/types.js";
 import type { MediaStrategyDeleteMultiple } from "@lucidcms/core/types";
 
-export default (
-	client: NodeJsClient<S3Client>,
-	pluginOptions: PluginOptions,
-) => {
+export default (client: S3Client, pluginOptions: PluginOptions) => {
 	const deleteMultiple: MediaStrategyDeleteMultiple = async (keys) => {
 		try {
 			const command = new DeleteObjectsCommand({