+++ packages/plugin-s3/src/services/delete-single.ts
@@ -1,13 +1,8 @@
-import T from "../translations/index.js";
 import { type S3Client, DeleteObjectCommand } from "@aws-sdk/client-s3";
-import type { NodeJsClient } from "@smithy/types";
 import type { PluginOptions } from "../types/types.js";
 import type { MediaStrategyDeleteSingle } from "@lucidcms/core/types";
 
-export default (
-	client: NodeJsClient<S3Client>,
-	pluginOptions: PluginOptions,
-) => {
+export default (client: S3Client, pluginOptions: PluginOptions) => {
 	const deletSingle: MediaStrategyDeleteSingle = async (key) => {
 		try {
 			const command = new DeleteObjectCommand({