+++ packages/plugin-s3/src/services/update-single.ts
@@ -1,13 +1,9 @@
 import { type S3Client, DeleteObjectCommand } from "@aws-sdk/client-s3";
-import type { NodeJsClient } from "@smithy/types";
 import type { PluginOptions } from "../types/types.js";
 import type { MediaStrategyUpdateSingle } from "@lucidcms/core/types";
 import uploadSingle from "./upload-single.js";
 
-export default (
-	client: NodeJsClient<S3Client>,
-	pluginOptions: PluginOptions,
-) => {
+export default (client: S3Client, pluginOptions: PluginOptions) => {
 	const updateSingle: MediaStrategyUpdateSingle = async (oldKey, props) => {
 		try {
 			const uploadRes = await uploadSingle(client, pluginOptions)(props);